Refactor video helpers into shared module
